requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
Pillow==10.4.0
django-environ==0.11.2
celery==5.4.0
redis==5.0.7
whitenoise==6.7.0
qrcode==7.4.2
python-dateutil==2.9.0
gunicorn==22.0.0
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
